Moved form components to Core/Commons/Form. Updated references and styling
